MVC for blog page
